requirements.txt
flask~=2.2.2
flask_cors
openai
langchain~=0.0.324
Flask-SocketIO
pypdf==3.16.1
tiktoken~=0.5.1
faiss-cpu
youtube-transcript-api
langchain-experimental
neo4j
wikipedia
python-dotenv~=0.21.0
duckduckgo-search
pandas==1.5.1
numpy==1.24.1
requests==2.28.1
sentence_transformers
scikit-learn~=1.3.0
pydantic~=2.4.2
apify-client